# (Further) Improve Bundler::Settings#[] performance and memory usage