--- a/bundler/lib/bundler/settings.rb
+++ b/bundler/lib/bundler/settings.rb
@@ -102,10 +102,13 @@ module Bundler
     def [](name)
       key = key_for(name)
 
-      values = configs.values
-      values.map! {|config| config[key] }
-      values.compact!
-      value = values.first
+      value = nil
+      configs.each do |_, config|
+        if config[key]
+          value = config[key]
+          break
+        end
+      end
 
       converted_value(value, name)
     end
@@ -316,7 +319,7 @@ module Bundler
     private
 
     def configs
-      {
+      @configs ||= {
         :temporary => @temporary,
         :local => @local_config,
         :env => @env_config,