@technicalpickles 2023-08-29
I'm also including the results of memory_profiler on our app boot, focused on memory_profiler. The main data point to note is the difference in memory allocated in `lib/bundler/settings.rb`.

Before: 
[boot-time-bundler-master.txt](https://github.com/rubygems/rubygems/files/12459222/boot-time-bundler-master.txt)

```
   2,332,050  /Users/josh.nichols/workspace/rubygems/bundler/lib/bundler/settings.rb
```

After: 
[boot-time-bundler-only-improvements-5.txt](https://github.com/rubygems/rubygems/files/12459223/boot-time-bundler-only-improvements-5.txt)

```
   1,065,298  /Users/josh.nichols/workspace/rubygems/bundler/lib/bundler/settings.rb
```

That is a 54.3% reduction in allocated memory 😍