@martinemde 2023-08-29
Based on the variety of test failure, the test must need that memoized state cleared between runs.