@technicalpickles 2023-08-29
> Btw why isn't there a "return result of first truthy block" on Enumerable. I often need it.

Yep, it is `detect`/`find`. I included an implementation in my benchmark using it, but it wasn't faster than an each block.

> tests must need that memoized state cleared between runs

Hmm, I will check it out. I think it's still an improvement w/o the memoization if it comes to it though.