@technicalpickles 2023-08-29
I tried reverting the memoization, and these are still failing, so it's something else. For reference, these are the failing ones:

```
 bin/rspec ./spec/commands/cache_spec.rb:200 ./spec/commands/cache_spec.rb:188 ./spec/commands/cache_spec.rb:371  ./spec/plugins/install_spec.rb:25 ./spec/plugins/install_spec.rb:283
```

I personally find some of the failures tough to parse in general especially when they are spawning new processes to do the work. If there are any protips, would love to hear them 😁