@martinemde 2023-08-29
> Yep, it is detect/find.

I'm wishing for `map_find {|(v)| result_of_this_is_returned(v) }`.

Anyway, I tried the code and I get the same errors. Then I tried putting both old and new in the same file. The keys it fails on are set to `false`! I looked at this many times and didn't see the difference between them, but of course the difference is that compact! doesn't throw out false values, but this PR does.