@indirect 2023-08-29
Does `flat_map` plus `break` or `return` get you the behavior you're looking for?