@technicalpickles 2023-08-30
> Does `flat_map` plus `break` or `return` get you the behavior you're looking for?

@indirect oh interesting, I hadn't considered using a `break` in the middle of a map 🤔  TIL you can break with a value, and it's interesting that is what ends up returned rather than the mapped thing.

I added this implementation to the benchmark:

```ruby
    key = key_for(name)
    value = configs.flat_map {|_, config| break config[key] unless config[key].nil? }

    converted_value(value, name)
```

Unfortunately, it uses more memory and is slow though 😞 See 'with_flat_map' on https://gist.github.com/technicalpickles/91933d34dcd47bfc99d6ac0e53fcdb1e